Updates gosigar to most recent version. Fixes build problem. Exclude memory also from solaris build
