+++ metricbeat/module/system/cpu/helper_test.go
@@ -1,8 +1,10 @@
 // +build !integration
+// +build darwin freebsd linux openbsd windows
 
 package cpu
 
 import (
+	"runtime"
 	"testing"
 
 	"github.com/elastic/gosigar"
@@ -60,4 +62,20 @@ func TestCpuPercentage(t *testing.T) {
 
 	assert.Equal(t, cpu2.UserPercent, 0.9502)
 	assert.Equal(t, cpu2.SystemPercent, 0.0448)
+}
+
+func TestGetSystemLoad(t *testing.T) {
+
+	if runtime.GOOS == "windows" {
+		return //no load data on windows
+	}
+
+	load, err := GetSystemLoad()
+
+	assert.NotNil(t, load)
+	assert.Nil(t, err)
+
+	assert.True(t, (load.Load1 > 0))
+	assert.True(t, (load.Load5 > 0))
+	assert.True(t, (load.Load15 > 0))
 }
