+++ metricbeat/module/system/filesystem/helper.go
@@ -1,3 +1,5 @@
+// +build darwin freebsd linux openbsd windows
+
 package filesystem
 
 import (
