+++ metricbeat/module/system/filesystem/helper_test.go
@@ -1,4 +1,5 @@
 // +build !integration
+// +build darwin freebsd linux openbsd windows
 
 package filesystem
